add bfs

+++ include/fs.h
@@ -1,4 +1,4 @@
 #ifndef __FS_H
 #define __FS_H
-uint32_t get_page_num();
+uint64_t get_page_num();
 #endif
